Big Bootloader Change
Structure change to the bootloaders.

All bootloaders now reference and use the version of LUFA that is in the
core directory. Removed individual directories for boards. You now
compile different versions of the bootloader by calling "make
<insertboardname>". This allows different button and LED layouts to be
changed without having to edit the source files.

Adding boards to the bootloader is easier now. Just open BootloaderCDC.h
and add a #elif defined for your board and then a make definition in the
makefile.

Reduced the size of the bootloader directory from 46MB to 276KB.